rename component home at index and merge
